+++ tests/testthat/test-create_fusen_rsproject.R
@@ -2,19 +2,19 @@
 dummypackage <- tempfile(pattern = "dummy")
 dir.create(dummypackage)
 
-raw_rmd_template <- readLines(system.file("dev-template-full.Rmd", package = "fusen"))
-expected_rmd_template_with_fusen_name <- gsub("<my_package_name>", "foosen", raw_rmd_template)
+expected_rmd_template_with_fusen_name <- readLines(system.file("flat-template-full.Rmd", package = "fusen")) %>%
+  gsub("<my_package_name>", "foosen", .) %>%
+  gsub("flat_template[.]Rmd", "flat_full.Rmd", .)
 
 withr::with_dir(dummypackage, {
   test_that("Create fusen", {
     path_foosen <- file.path(dummypackage, "foosen")
-    path_dev_history <- file.path(path_foosen, "dev", "dev_history.Rmd")
-    create_fusen(path_foosen, name = "full", open = FALSE)
+    path_dev_history <- suppressMessages(create_fusen(path_foosen, template = "full", open = FALSE))
 
     expect_true(dir.exists(path_foosen))
-    expect_true(file.exists(path_dev_history))
+    expect_true(all(file.exists(path_dev_history)))
 
-    actual_dev_history <- readLines(path_dev_history)
+    actual_dev_history <- readLines(path_dev_history[grepl("flat", path_dev_history)])
     expect_identical(
       actual_dev_history,
       expected_rmd_template_with_fusen_name
@@ -31,14 +31,14 @@ withr::with_dir(dummypackage, {
     # Can not overwrite
     expect_error(
       expect_message(
-        create_fusen(path_foosen, name = "full", open = FALSE),
+        create_fusen(path_foosen, template = "full", open = FALSE),
         "Could not create fusen project"
       )
     )
 
     # Can overwrite with message
     expect_message(
-      create_fusen(path_foosen, name = "full", open = FALSE, overwrite = TRUE),
+      create_fusen(path_foosen, template = "full", open = FALSE, overwrite = TRUE),
       "Some files may be overwritten"
     )
 
@@ -50,15 +50,16 @@ withr::with_dir(dummypackage, {
     }
   })
 })
-
+# clean
+unlink(dummypackage, recursive = TRUE)
 
 ## Create in a subdirectory ----
 dummysubdir <- tempfile(pattern = "subdir/subdir2/dummy")
 test_that("Can create in a subdirectory", {
-  expect_error(create_fusen(dummysubdir, name = "full", open = FALSE), regexp = NA)
+  expect_error(suppressMessages(create_fusen(dummysubdir, template = "full", open = FALSE)), regexp = NA)
   expect_true(dir.exists(dummysubdir))
 })
-
+unlink(dummysubdir, recursive = TRUE)
 
 ## Test gui ----
 
@@ -75,12 +76,14 @@ create_dummygui <- function() {
 dummygui <- create_dummygui()
 withr::with_dir(dummygui$dirname, {
   test_that("Can create in a subdirectory", {
-    expect_error(
-      create_fusen_gui(dummygui$basename, name = "teaching", with_git = FALSE),
+    dev_path <- expect_error(
+      suppressMessages(
+        create_fusen_gui(dummygui$basename, template = "teaching", with_git = FALSE)
+      ),
       regexp = NA # expect no errors
     )
     expect_true(dir.exists(dummygui$path))
-    expect_true(file.exists(file.path(dummygui$path, "dev", "dev_history.Rmd")))
+    expect_true(file.exists(dev_path))
     expect_false(dir.exists(file.path(dummygui$path, ".git/")))
   })
 })
@@ -90,36 +93,35 @@ unlink(dummygui$path, recursive = TRUE)
 ## Test initialise git ----
 
 # In cli
+dummypackage <- tempfile(pattern = "dummy")
+dir.create(dummypackage)
 withr::with_dir(dummypackage, {
   test_that("Create a fusen project with git at the cli", {
     path_gigit <- file.path(dummypackage, "gigit")
-    path_dev_history <- file.path(path_gigit, "dev", "dev_history.Rmd")
-
-    create_fusen(path_gigit, name = "full", open = FALSE, with_git = TRUE)
+    dev_path <- suppressMessages(create_fusen(path_gigit, template = "full", open = FALSE, with_git = TRUE))
 
     expect_true(dir.exists(path_gigit))
-    expect_true(file.exists(path_dev_history))
-    # git is initialised
+    expect_true(all(file.exists(dev_path)))
+    # git is initialized
     expect_true(dir.exists(file.path(path_gigit, ".git")))
   })
 })
+unlink(dummypackage, recursive = TRUE)
 
 # In Rstudio GUI project wizard
 dummygui <- create_dummygui()
 withr::with_dir(dummygui$dirname, {
   test_that("Create a fusen project with git using Rstudio GUI", {
-    expect_error(
-      create_fusen_gui(dummygui$basename, name = "teaching", with_git = TRUE),
+    dev_file <- expect_error(
+      suppressMessages(
+        create_fusen_gui(dummygui$basename, template = "teaching", with_git = TRUE)
+      ),
       regexp = NA # expect no errors
     )
     expect_true(dir.exists(dummygui$path))
-    expect_true(file.exists(file.path(dummygui$path, "dev", "dev_history.Rmd")))
-    # git is initialised
+    expect_true(file.exists(dev_file))
+    # git is initialized
     expect_true(dir.exists(file.path(dummygui$path, ".git/")))
   })
 })
 unlink(dummygui$path, recursive = TRUE)
-
-
-unlink(dummypackage, recursive = TRUE)
-unlink(dummysubdir, recursive = TRUE)