+++ tests/testthat/test-inflate.R
@@ -1,21 +1,24 @@
 # Create a new project
-dummypackage <- tempfile("dummypackage")
+dummypackage <- tempfile("inflate.tests")
 dir.create(dummypackage)
 
 # {fusen} steps
 fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
-dev_file <- add_dev_history(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+dev_file <- suppressMessages(add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE))
+flat_file <- dev_file[grepl("flat_", dev_file)]
 
 usethis::with_project(dummypackage, {
-
   # More complicated example for tests
   testfile <- "tests-templates/dev-template-tests.Rmd"
   file.copy(
     system.file(testfile, package = "fusen"),
-    dev_file,
+    flat_file,
     overwrite = TRUE
   )
-  inflate(pkg = dummypackage, rmd = dev_file, name = "Get started", check = FALSE)
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file,
+            name = "Get started", check = FALSE)
+  )
 
   test_that("inflate() worked correctly", {
     # R files
@@ -68,11 +71,16 @@ usethis::with_project(dummypackage, {
 
     # _no roxygen but examples
     my_norox2_lines <- readLines(my_norox2functionfile)
-    expect_true(all(my_norox2_lines == c(
-      "#' @noRd", "#' @examples", "#' \\dontrun{",
+    expect_equal(my_norox2_lines, c(
+      "#' @noRd", "#' @examples",
+      "#' \\dontrun{", "#' # comment",
       "#' my_norox2(10)", "#' }",
+      "#' ",
+      "#' \\dontrun{",
+      "#' # comment",
+      "#' my_norox2(12)", "#' }",
       "my_norox2 <- function(x) {", "  x + 10", "}"
-    )))
+    ))
     # _extra empty line and examples
     my_space_lines <- readLines(my_spacefunctionfile)
     expect_true(all(my_space_lines[6:10] == c(
@@ -157,18 +165,22 @@ if (packageVersion("parsermd") > "0.1.2") {
 
   # {fusen} steps
   fill_description(pkg = dummypackage.special, fields = list(Title = "Dummy Package"))
-  dev_file <- add_dev_history(pkg = dummypackage.special, overwrite = TRUE, open = FALSE)
+  dev_file <- add_flat_template(pkg = dummypackage.special, overwrite = TRUE, open = FALSE)
+  flat_file <- dev_file[grepl("flat_", dev_file)]
 
   usethis::with_project(dummypackage.special, {
 
     testfile <- "tests-templates/dev-template-tests-special-char.Rmd"
     file.copy(
       system.file(testfile, package = "fusen"),
-      dev_file,
+      flat_file,
       overwrite = TRUE
     )
 
-    inflate(pkg = dummypackage.special, rmd = dev_file, name = "Get started", check = FALSE)
+    suppressMessages(
+      inflate(pkg = dummypackage.special, rmd = flat_file,
+              name = "Get started", check = FALSE)
+    )
 
     test_that("inflate with special yaml worked correctly", {
       # R files
@@ -181,12 +193,19 @@ if (packageVersion("parsermd") > "0.1.2") {
 # Test no attachment and no check when asked ----
 unlink(file.path(dummypackage, "DESCRIPTION"), recursive = TRUE)
 fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
-dev_file <- add_dev_history(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+dev_file <- add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+flat_file <- dev_file[grepl("flat_", dev_file)]
 
 usethis::with_project(dummypackage, {
-  inflate(pkg = dummypackage, rmd = dev_file, name = "Get started", check = FALSE, document = FALSE)
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file, name = "Get started",
+            check = FALSE, document = FALSE)
+  )
   desc_lines <- readLines(file.path(dummypackage, "DESCRIPTION"))
-  expect_false("Imports:" %in% desc_lines)
+
+  test_that("no attachment run", {
+    expect_false("Imports:" %in% desc_lines)
+  })
 
   # Clean R, tests and vignettes
   unlink(file.path(dummypackage, "R"), recursive = TRUE)
@@ -199,10 +218,14 @@ usethis::with_project(dummypackage, {
 usethis::with_project(dummypackage, {
   file.copy(
     system.file("tests-templates/dev-template-no-example-no-tests.Rmd", package = "fusen"),
-    dev_file,
+    flat_file,
     overwrite = TRUE
   )
-  inflate(pkg = dummypackage, rmd = dev_file, name = "Get started", check = FALSE)
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file,
+            name = "Get started", check = FALSE)
+  )
+
   test_that("inflate() output no error", {
     expect_true(file.exists(file.path(dummypackage, "vignettes", "get-started.Rmd")))
     expect_true(file.exists(file.path(dummypackage, "R", "my_median.R")))
@@ -218,11 +241,16 @@ usethis::with_project(dummypackage, {
 usethis::with_project(dummypackage, {
   file.copy(
     system.file("tests-templates/dev-template-test-parse-nothing.Rmd", package = "fusen"),
-    dev_file,
+    flat_file,
     overwrite = TRUE
   )
   test_that("inflate() output message", {
-    expect_message(inflate(pkg = dummypackage, rmd = dev_file, name = "Get started", check = FALSE))
+    suppressMessages(
+      expect_message(
+        inflate(pkg = dummypackage, rmd = flat_file,
+                name = "Get started", check = FALSE)
+      )
+    )
   })
   # Clean R, tests and vignettes
   unlink(file.path(dummypackage, "R"), recursive = TRUE)
@@ -232,20 +260,32 @@ usethis::with_project(dummypackage, {
 
 # Tests errors - vignette already exists ----
 usethis::with_project(dummypackage, {
-
-  inflate(pkg = dummypackage, rmd = dev_file, name = "Get started",
-          check = FALSE, overwrite = "yes")
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file,
+            name = "Get started",
+            check = FALSE, overwrite = "yes")
+  )
 
   test_that("inflate() output error when second time (not interactive)", {
-    expect_error(inflate(pkg = dummypackage, rmd = dev_file, name = "Get started",
-                         check = FALSE))
-    expect_error(inflate(pkg = dummypackage, rmd = dev_file, name = "Get started",
-                         check = FALSE, overwrite = 'no'))
+    expect_error(
+      suppressMessages(
+        inflate(pkg = dummypackage, rmd = flat_file,
+                name = "Get started",
+                check = FALSE))
+    )
+    expect_error(
+      suppressMessages(
+        inflate(pkg = dummypackage, rmd = flat_file,
+                name = "Get started",
+                check = FALSE, overwrite = 'no'))
+    )
   })
 
   # No error with overwrite = 'yes'
-  inflate(pkg = dummypackage, rmd = dev_file, name = "Get started",
-          check = FALSE, overwrite = "yes")
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file, name = "Get started",
+            check = FALSE, overwrite = "yes")
+  )
 
   test_that("inflate() output no error", {
     expect_true(file.exists(file.path(dummypackage, "vignettes", "get-started.Rmd")))
@@ -261,11 +301,15 @@ usethis::with_project(dummypackage, {
 usethis::with_project(dummypackage, {
   file.copy(
     system.file("tests-templates/dev-template-stop-duplicate-fun.Rmd", package = "fusen"),
-    dev_file,
+    flat_file,
     overwrite = TRUE
   )
   test_that("inflate() output error duplicate functions", {
-    expect_error(inflate(pkg = dummypackage, rmd = dev_file, name = "Get started", check = FALSE))
+    expect_error(
+      suppressMessages(
+        inflate(pkg = dummypackage, rmd = flat_file,
+                name = "Get started", check = FALSE))
+    )
   })
   # Clean R, tests and vignettes
   unlink(file.path(dummypackage, "R"), recursive = TRUE)
@@ -275,11 +319,15 @@ usethis::with_project(dummypackage, {
   # Tests errors - duplicate chunk names
   file.copy(
     system.file("tests-templates/dev-template-stop-duplicate-label.Rmd", package = "fusen"),
-    dev_file,
+    flat_file,
     overwrite = TRUE
   )
   test_that("inflate() output error duplicate label names for vignette", {
-    expect_error(inflate(pkg = dummypackage, rmd = dev_file, name = "Get started", check = FALSE))
+    expect_error(
+      suppressMessages(
+        inflate(pkg = dummypackage, rmd = flat_file, name = "Get started", check = FALSE)
+      )
+    )
   })
   # Clean R, tests and vignettes
   unlink(file.path(dummypackage, "R"), recursive = TRUE)
@@ -295,11 +343,14 @@ usethis::with_project(dummypackage, {
 
   # Add
   # {fusen} steps
-  dev_file <- add_dev_history(pkg = dummypackage, overwrite = TRUE, open = FALSE)
-  inflate(pkg = dummypackage, rmd = dev_file, name = "Get started", check = FALSE)
+  dev_file <- add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file,
+            name = "Get started", check = FALSE)
+  )
 
-  test_that("add_dev_history inflates with .Rproj and no .here", {
-    expect_true(file.exists(dev_file))
+  test_that("add_flat_template inflates with .Rproj and no .here", {
+    expect_true(file.exists(flat_file))
     expect_false(file.exists(file.path(dummypackage, ".here")))
 
     rbuildignore_file <- file.path(dummypackage, ".Rbuildignore")
@@ -327,7 +378,9 @@ usethis::with_project(dummypackage, {
 
 # Test no errors - clean name for vignette ----
 usethis::with_project(dummypackage, {
-  inflate(pkg = dummypackage, rmd = dev_file, name = "# y  _ p n@ \u00E9 ! 1", check = FALSE)
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file, name = "# y  _ p n@ \u00E9 ! 1", check = FALSE)
+  )
   # Vignette name is also cleaned by {usethis} for special characters
   vignette_path <- file.path(dummypackage, "vignettes", "y-p-n---1.Rmd")
 
@@ -346,7 +399,9 @@ usethis::with_project(dummypackage, {
 })
 
 usethis::with_project(dummypackage, {
-  inflate(pkg = dummypackage, rmd = dev_file, name = "# y  _ p n@ \u00E9 ! 1", check = FALSE)
+  suppressMessages(
+    inflate(pkg = dummypackage, rmd = flat_file, name = "# y  _ p n@ \u00E9 ! 1", check = FALSE)
+  )
   # Vignette name is also cleaned by {usethis} for special characters
   vignette_path <- file.path(dummypackage, "vignettes", "y-p-n---1.Rmd")
 
@@ -370,11 +425,16 @@ unlink(dummypackage, recursive = TRUE)
 # Test stop when no DESCRIPTION file ----
 dummypackage <- tempfile("descpackage")
 dir.create(dummypackage)
-dev_file <- add_dev_history(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+dev_file <- add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+flat_file <- dev_file[grepl("flat_", dev_file)]
 
 usethis::with_project(dummypackage, {
   test_that("stop when no DESCRIPTION file", {
-    expect_error(inflate(pkg = dummypackage, rmd = dev_file, check = FALSE), "DESCRIPTION file")
+    expect_error(
+      suppressMessages(
+        inflate(pkg = dummypackage, rmd = flat_file, check = FALSE)
+      ),
+      "DESCRIPTION file")
   })
 })
 
@@ -390,19 +450,24 @@ if (exists("dummypackage.special")) {
 
 dummypackage <- tempfile("nordpackage")
 dir.create(dummypackage)
-dev_file <- add_dev_history(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+dev_file <- add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE)
+flat_file <- dev_file[grepl("flat_", dev_file)]
 fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
 
 usethis::with_project(dummypackage, {
   file.copy(
     system.file("tests-templates/dev-template-nord-but-example.Rmd", package = "fusen"),
-    dev_file,
+    flat_file,
     overwrite = TRUE
   )
 
   test_that("Deal with noRd but examples", {
     # No error
-    expect_error(inflate(pkg = dummypackage, rmd = dev_file, check = FALSE), regexp = NA)
+    expect_error(
+      suppressMessages(
+        inflate(pkg = dummypackage, rmd = flat_file, check = FALSE)
+      ),
+      regexp = NA)
     # Check error
     skip_on_os(os = c("windows", "solaris"))
 
@@ -416,7 +481,7 @@ usethis::with_project(dummypackage, {
 unlink(dummypackage, recursive = TRUE)
 
 # Test checks all templates with inflate dots (...) ----
-alltemp <- tempfile("all_templates")
+alltemp <- tempfile("all.templates.inflate")
 dir.create(alltemp)
 
 for (pkgname in c("full", "teaching")) {
@@ -424,42 +489,45 @@ for (pkgname in c("full", "teaching")) {
   # No "additional" with create_fusen
   # {fusen} steps
   path_foosen <- file.path(alltemp, pkgname)
-  create_fusen(path_foosen, name = pkgname, open = FALSE)
-  dev_file <- file.path(path_foosen, "dev", "dev_history.Rmd")
+  dev_file <- create_fusen(path_foosen, template = pkgname, open = FALSE)
+  flat_file <- dev_file[grepl("flat_", dev_file)]
 
   usethis::with_project(path_foosen, {
-    # withr::with_package("MASS", {
+# browser()
+    # Do not check inside check if on CRAN
+    skip_on_os(os = c("windows", "solaris"))
 
-      # Do not check inside check if on CRAN
-      skip_on_os(os = c("windows", "solaris"))
+    fill_description(pkg = path_foosen, fields = list(Title = "Dummy Package"))
+    usethis::use_gpl_license()
 
-      fill_description(pkg = path_foosen, fields = list(Title = "Dummy Package"))
-      usethis::use_gpl_license()
-
-      test_that(paste("Check returns OK for template", pkgname), {
-        # No redirection of stdout/stderr when non-interactive
-        expect_error(
-          inflate(pkg = path_foosen, rmd = dev_file, name = "exploration",
+    test_that(paste("Check returns OK for template", pkgname), {
+      # No redirection of stdout/stderr when non-interactive
+      expect_error(
+        suppressMessages(
+          inflate(pkg = path_foosen, rmd = flat_file, name = "exploration",
                   check = TRUE, quiet = TRUE, args = c("--no-manual"),
-                  overwrite = TRUE),
-          regexp = NA)
+                  overwrite = TRUE)
+        ),
+        regexp = NA)
 
-        skip_if_not(interactive())
-        # Needs MASS, lattice, Matrix installed
-        # quiet and checkdir
-        checkdir <- file.path(alltemp, paste0("checkout", pkgname))
-        expect_error(
-          inflate(pkg = path_foosen, rmd = dev_file, name = "exploration",
+      skip_if_not(interactive())
+      # Needs MASS, lattice, Matrix installed
+      # quiet and checkdir
+      checkdir <- file.path(alltemp, paste0("checkout", pkgname))
+      expect_error(
+        suppressMessages(
+          inflate(pkg = path_foosen, rmd = flat_file, name = "exploration",
                   check = TRUE, check_dir = checkdir, quiet = TRUE,
                   args = c("--no-manual"),
-                  overwrite = TRUE),
-          regexp = NA)
+                  overwrite = TRUE)
+        ),
+        regexp = NA)
 
-        # Should not be any errors with templates
-        check_lines <- readLines(file.path(checkdir, paste0(basename(path_foosen), ".Rcheck"), "00check.log"))
-        expect_equal(check_lines[length(check_lines)], "Status: OK")
-        unlink(checkdir, recursive = TRUE)
-      })
+      # Should not be any errors with templates
+      check_lines <- readLines(file.path(checkdir, paste0(basename(path_foosen), ".Rcheck"), "00check.log"))
+      expect_equal(check_lines[length(check_lines)], "Status: OK")
+      unlink(checkdir, recursive = TRUE)
+    })
     # })
   })