+++ NEWS.md
@@ -1,7 +1,13 @@
 # fusen 0.2.4.9000
 
+## Breaking changes
+
+* `add_flat_template()` superseeds `add_dev_history()` with more advanced possibilities
+* `add_dev_history()` is deprecated
+
 ## Major changes
 
+* Improve documentation included in flat templates to reflect changes in using dev_history file
 * Add Rstudio Addin to insert chunks for new function (@ColinFay)
 * Deal with `\dontrun{}` in example chunks
 * Allow short names for chunks: dev, fun, ex, test